Namespace the service classes
